Changes: RunOnUiThread in ISensorActivity for UICallbacks (Thread Conflict) Changes: Added startReadISensor to ISensor Changes: UI Button Disconnect to Stop Reading.

+++ app/src/main/java/com/example/bluetoothinterface/interfaces/IQMSensor.java
@@ -10,6 +10,6 @@ import java.util.List;
  */
 
 public interface IQMSensor {
-    double lostFrames(List<DataFrameFactory> last5SecondsData);
+    int lostFrames(List<DataFrameFactory> last5SecondsData);
     boolean shouldDisconnect(Date lastReadTime);
 }
